htable: track the number of entries.
This is useful to check how many entries with the same name are
available when there is a collision.
We rely on the fact that our hash function is awesome and collission
happens only when the keys are same. We are in a pickle if two very
different keys generate the same hash, which is a possibility.